# docs/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Academic Knowledge Backend",
    "description": "Backend API for course-centric academic knowledge management.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/courses": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "List Courses Endpoint",
        "operationId": "list_courses_endpoint_api_v1_courses_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "default": true,
              "title": "Is Active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseRead"
                  },
                  "title": "Response List Courses Endpoint Api V1 Courses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Create Course Endpoint",
        "operationId": "create_course_endpoint_api_v1_courses_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/review/queue": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Get Review Queue",
        "description": "List documents in NEEDS_REVIEW visible to the current user.",
        "operationId": "get_review_queue_api_v1_review_queue_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewQueueItem"
                  },
                  "type": "array",
                  "title": "Response Get Review Queue Api V1 Review Queue Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/review/{document_id}": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Get Review Detail",
        "description": "Get full review interface data for a document.",
        "operationId": "get_review_detail_api_v1_review__document_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/review/{document_id}/decide": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Decide Review",
        "description": "Submit a review decision for a document.",
        "operationId": "decide_review_api_v1_review__document_id__decide_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecideRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDecisionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/courses/{course_code}/mindmap": {
      "get": {
        "tags": [
          "mindmap"
        ],
        "summary": "Get Mindmap",
        "operationId": "get_mindmap_api_v1_courses__course_code__mindmap_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MindmapGenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/courses/{course_code}/mindmap/generate": {
      "post": {
        "tags": [
          "mindmap"
        ],
        "summary": "Generate Mindmap",
        "operationId": "generate_mindmap_api_v1_courses__course_code__mindmap_generate_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MindmapGenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CourseCreate": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 20,
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "topic_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Summary"
          },
          "short_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Description"
          },
          "review_sla_hours": {
            "type": "integer",
            "maximum": 72.0,
            "minimum": 24.0,
            "title": "Review Sla Hours",
            "default": 48
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "CourseCreate"
      },
      "CourseRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "topic_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Summary"
          },
          "short_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Description"
          },
          "topic_tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topic Tags"
          },
          "review_sla_hours": {
            "type": "integer",
            "title": "Review Sla Hours"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "description",
          "topic_summary",
          "short_description",
          "topic_tags",
          "review_sla_hours",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "CourseRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MindmapGenerateResponse": {
        "properties": {
          "course_code": {
            "type": "string",
            "title": "Course Code"
          },
          "is_cached": {
            "type": "boolean",
            "title": "Is Cached"
          },
          "concept_graph": {
            "additionalProperties": {
              "items": {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Concept Graph"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "course_code",
          "is_cached",
          "concept_graph",
          "generated_at"
        ],
        "title": "MindmapGenerateResponse"
      },
...
      "ReviewDecideRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "final_contribution_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Contribution Type"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "ReviewDecideRequest"
      },
      "ReviewDecisionRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "final_contribution_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Contribution Type"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "decided_at": {
            "type": "string",
            "format": "date-time",
            "title": "Decided At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "decision",
          "final_contribution_type",
          "note",
          "decided_at"
        ],
        "title": "ReviewDecisionRead"
      },
      "ReviewDetailResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "course_code": {
            "type": "string",
            "title": "Course Code"
          },
          "document_tier": {
            "type": "string",
            "title": "Document Tier"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "evaluation_report": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Report"
          },
          "state_logs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "State Logs"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "course_code",
          "document_tier",
          "status",
          "evaluation_report",
          "state_logs"
        ],
        "title": "ReviewDetailResponse"
      },
      "ReviewQueueItem": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "course_code": {
            "type": "string",
            "title": "Course Code"
          },
          "document_tier": {
            "type": "string",
            "title": "Document Tier"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "sla_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Deadline"
          },
          "sla_breached": {
            "type": "boolean",
            "title": "Sla Breached"
          },
          "no_reviewer_flag": {
            "type": "boolean",
            "title": "No Reviewer Flag"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "course_code",
          "document_tier",
          "status",
          "sla_deadline",
          "sla_breached",
          "no_reviewer_flag",
          "uploaded_at"
        ],
        "title": "ReviewQueueItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}